+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/amqp/remoting/codec/AmqpCodeDecoder.java
@@ -19,7 +19,7 @@ package org.apache.eventmesh.runtime.core.protocol.amqp.remoting.codec;
 
 import io.netty.util.Attribute;
 import org.apache.eventmesh.runtime.core.protocol.amqp.exception.MalformedFrameException;
-import org.apache.eventmesh.runtime.core.protocol.amqp.remoting.AMQPFrame;
+import org.apache.eventmesh.runtime.core.protocol.amqp.remoting.frame.AMQPFrame;
 import org.apache.eventmesh.runtime.core.protocol.amqp.remoting.constants.ProtocolKey;
 import org.apache.eventmesh.runtime.core.protocol.amqp.remoting.protocol.ProtocolFrame;
 import org.apache.eventmesh.runtime.core.protocol.amqp.remoting.protocol.ProtocolVersion;
