+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/amqp/remoting/frame/AMQData.java
@@ -1,12 +1,31 @@
-package org.apache.eventmesh.runtime.core.protocol.amqp.remoting.frame;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import io.netty.buffer.ByteBuf;
+package org.apache.eventmesh.runtime.core.protocol.amqp.remoting.frame;
 
 import java.io.Serializable;
 
+import io.netty.buffer.ByteBuf;
+
+/**
+ *
+ */
 public interface AMQData extends Serializable {
 
     void encode(ByteBuf buf);
 
-    void recycle();
 }
